tests/crud_comprehensive/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
asyncpg>=0.28.0
PyJWT>=2.8.0
cryptography>=41.0.0
pydantic>=2.0.0
faker>=19.0.0
python-dotenv>=1.0.0
tavern[pytest]>=2.16.0

# CI/CD and Performance Testing
pytest-xdist>=3.3.1          # Parallel test execution
pytest-html>=4.1.1           # HTML test reports
pytest-json-report>=1.5.0    # JSON test results
pytest-timeout>=2.1.0        # Test timeout management
pytest-benchmark>=4.0.0      # Performance benchmarking
pytest-cov>=4.1.0           # Code coverage
pytest-mock>=3.11.1         # Mocking utilities
pytest-retry>=1.6.1         # Retry flaky tests

# Database Connection
psycopg2-binary>=2.9.0     # PostgreSQL adapter